+++ lua/neg/groups/plugins/alpha.lua
@@ -1,0 +1,10 @@
+local p = require('neg.palette')
+
+-- alpha-nvim start screen
+return {
+  AlphaHeader   = { fg = p.keyword3_color },
+  AlphaFooter   = { fg = p.comment_color },
+  AlphaButtons  = { fg = p.default_color },
+  AlphaShortcut = { fg = p.search_color },
+}
+