+++ lua/neg/groups/plugins/treesitter_playground.lua
@@ -1,0 +1,10 @@
+local p = require('neg.palette')
+
+-- nvim-treesitter/playground (best-effort)
+return {
+  TSPlaygroundFocus   = { fg = p.search_color, bold = true },
+  TSPlaygroundLang    = { fg = p.preproc_light_color },
+  TSPlaygroundQuery   = { fg = p.literal1_color },
+  TSPlaygroundHelp    = { fg = p.comment_color },
+}
+